+++ atmsim/atm_interface.h
@@ -48,7 +48,7 @@ class ATMInterface{
     void draw_gui(std::string text);
     void action_parser(std::string text);
 
-    std::string alt_to_string(double value);
+    std::string alt_to_string(float value);
     // std::string alt_display();
 
     public:
