+++ atmsim/utils.cpp
@@ -3,17 +3,17 @@
 
 namespace Utils{
 
-    double calculate_angle(Eigen::Vector3d p1, Eigen::Vector3d p2){
+    float calculate_angle(Eigen::Vector3d p1, Eigen::Vector3d p2){
         return  p1[1]-p2[1] < 0 ? atan((p1[0]-p2[0])/
                     (p1[1]-p2[1])) 
                     : PI + atan((p1[0]-p2[0])/
                     (p1[1]-p2[1]));
     }
 
-    double rad_to_deg(double value){
+    float rad_to_deg(float value){
         return (value*180)/PI;
     }
-    double deg_to_rad(double value){
+    float deg_to_rad(float value){
         return value *(PI/180);
     }
